requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cftime==1.6.2
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.1.1
fastapi==0.98.0
fonttools==4.43.1
frozenlist==1.3.3
google-search-results==2.4.2
google-serp-api==1.0.4
h11==0.14.0
httptools==0.5.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
multidict==6.0.4
netCDF4==1.6.4
numpy==1.26.0
openai==0.27.8
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pydantic==1.10.9
pydantic_core==2.10.1
pygmt==0.10.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
requests==2.31.0
serpapi==0.1.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
xarray==2023.9.0
yarl==1.9.2
